condition checked before dropdatabase in test cases

+++ src/test/java/analytics/bolt/ParsingBoltSYWMockTest.java
@@ -84,7 +84,10 @@ public class ParsingBoltSYWMockTest {
 	
 	@AfterClass
 	public static void cleanUp(){
-		db.dropDatabase();
+		if(db.toString().equalsIgnoreCase("FongoDB.test"))
+			   db.dropDatabase();
+			  else
+			   Assert.fail("Something went wrong. Tests connected to " + db.toString());
 	}
 
 }
