+++ src/test/java/analytics/bolt/PersistBoltOccassionTest.java
@@ -128,7 +128,10 @@ public class PersistBoltOccassionTest {
 
 	@AfterClass
 	public static void teardown() {
-		db.dropDatabase();
+		if(db.toString().equalsIgnoreCase("FongoDB.test"))
+			   db.dropDatabase();
+			  else
+			   Assert.fail("Something went wrong. Tests connected to " + db.toString());
 	}
 
 }
