+++ src/test/java/analytics/bolt/SocialBoltMockTest.java
@@ -84,7 +84,10 @@ public class SocialBoltMockTest {
 	}
 	@AfterClass
 	public static void cleanUp(){
-		conn.dropDatabase();
+		if(conn.toString().equalsIgnoreCase("FongoDB.test"))
+			   conn.dropDatabase();
+			  else
+			   Assert.fail("Something went wrong. Tests connected to " + conn.toString());
 	}
 	
 }
