+++ src/test/java/analytics/util/StrategyMapperTest.java
@@ -41,7 +41,10 @@ public class StrategyMapperTest {
 	
 	@AfterClass
 	public static void cleanUp(){
-		db.dropDatabase();
+		if(db.toString().equalsIgnoreCase("FongoDB.test"))
+			   db.dropDatabase();
+			  else
+			   Assert.fail("Something went wrong. Tests connected to " + db.toString());
 	}
 	
 	@Test
